Make docker always do its checkout
The many renames are because something in the supplied test code goes
crazy with an ISO date in the path. So I have the actual maven build
with a fixed directory name. But the other pieces use the ISO date so
that they don't get cached. An even better idea would be to use the
repository sha for the "unique value" but that wasn't easy to do
without a repo checkout.